Initial_Push
Fanfic downloader is able to download story from URL through command line. Saving the file in a directory called "downloads" in the same src file. The story is saved in a folder of respective name, containing the txt files of the description and the chapters along with the zip of the story directory.